Welcome component with API keys added and working nicely
